+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.5.3] - 2023-04-03
 ### Added
 - the `deleteFolder` function now exists.
 - the `deleteArchiveRecord` function now exists.
@@ -60,7 +62,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The `Archive` type.
 - The `ClientConfig` type which allows an sdk user to specify their credentials and other configuration settings.
 
-[Unreleased]: https://github.com/permanentorg/permanent-sdk/base/compare/v0.5.2...HEAD
+[Unreleased]: https://github.com/permanentorg/permanent-sdk/base/compare/v0.5.3...HEAD
+[0.5.3]: https://github.com/permanentorg/permanent-sdk/base/releases/tag/v0.5.3
 [0.5.2]: https://github.com/permanentorg/permanent-sdk/base/releases/tag/v0.5.2
 [0.5.1]: https://github.com/permanentorg/permanent-sdk/base/releases/tag/v0.5.1
 [0.5.0]: https://github.com/permanentorg/permanent-sdk/base/releases/tag/v0.5.0